Separating atomic grids
